Give the Arduino-ESP32 profile ONE home, and let `labwired test` use it
The profile lived inside `snapshot capture` only. That was not a tidiness
problem: `labwired test` — the declarative runner that owns `stimuli:`,
assertions, result.json and JUnit, i.e. the one a USER writes tests
against — had none of it, so a classic-ESP32 Arduino sketch did not boot
there at all. It produced ~47 bytes of UART and stopped at max_steps. The
runner that could inject inputs could not boot the firmware; the runner
that could boot it could not inject. Ryan's rig needs both.

The setup now lives in core/system/xtensa/arduino_esp32_profile.rs and both
CLI commands call it. It is in `core` rather than the CLI so the WASM
bridge can converge on it too, and so a fix to the boot path cannot land
for one runner and silently miss another.

`core` gained no ELF-format dependency: the symbol map is resolved by the
caller (labwired-loader) and passed in. Reading object files is the
loader's job; deciding what to do with the addresses is this module's.

Boot profile is opt-in and NAMED in the script (`inputs.profile:
arduino-esp32`). The faithful thunk-free path stays the default, so a run
that took the fast boot says so rather than leaving a reader to infer which
of two boot paths produced a result.

Also here:
  * The fast boot gets the rom-boot step ceiling. Arduino bring-up plus a
    polling sketch does not fit in 50M steps, and at that ceiling the run
    stopped with 47 bytes and stop_reason=max_steps — which reads as broken
    firmware rather than as a budget the runner declined to grant.
  * `install_xtensa_freertos_workarounds` is skipped under the profile:
    both installed the SAME xthal_window_spill_nw thunk, and patching BREAK
    bytes over an already-patched site loses the saved original. One owner
    per thunk.
  * The extracted module prints nothing. It is a library; the thunk count
    comes back in `thunks_installed` so each runner phrases its own line,
    instead of every runner claiming to be "labwired-cli snapshot".

Verified: `snapshot capture` is unchanged — Ryan's lab is 12/12 PASS with
UART byte-identical to a binary built from main. (The .lwrs blob differs,
but it differs between two runs of the SAME binary at the same offset, so
it is not a byte-comparable artifact.)

NOT yet fixed, measured and recorded in the task: `labwired test` boots to
setup() but produces no loop() output, where `snapshot capture` at an equal
105M cycles produces 152 STATE lines. The remaining difference is per-step,
not setup: snapshot's loop maintains an IPI bridge (DPORT FROM_CPU mapped
to a CPU interrupt edge each cycle) that the generic loop does not.
